+++ src/components/ConnectWallet.tsx
@@ -4,32 +4,15 @@ import { BeaconWallet } from "@taquito/beacon-wallet";
 import { NetworkType, BeaconEvent, defaultEventCallbacks } from "@airgap/beacon-sdk";
 import Button from "./Button";
 import { WalletContext } from "./context/WalletContext";
-import { HoundInfo, HoundRarity } from "../scripts/hound-genome";
 import { TzContext } from "./context/TzToolKitContext";
 
 type ButtonProps = {
   Tezos: TezosToolkit;
-  hounds: Map<string, HoundInfo[]>;
   setPublicToken: Dispatch<SetStateAction<string | null>>;
   setWalletConnected: Dispatch<SetStateAction<boolean>>;
 };
 
-function addHoundToMap(map: Map<string, HoundInfo[]>, key: string, value: HoundInfo) {
-  if (!map.has(key)) map.set(key, []);
-  (map.get(key) as HoundInfo[]).push(value);
-}
-
-const generateHound = (genome: string, generation: number): HoundInfo => {
-  return {
-    nick: "Hound",
-    generation: generation,
-    id: Math.ceil(1241 + Math.random() * 2000),
-    genome: genome,
-    rarity: HoundRarity.COMMON,
-  };
-};
-
-function ConnectButton({ hounds, setPublicToken, setWalletConnected }: ButtonProps): JSX.Element {
+function ConnectButton({setPublicToken, setWalletConnected }: ButtonProps): JSX.Element {
   const walletInfo = useContext(WalletContext);
   const tzContext = useContext(TzContext);
 
@@ -79,21 +62,6 @@ function ConnectButton({ hounds, setPublicToken, setWalletConnected }: ButtonPro
       if (activeAccount) {
         const userAddress = await wallet.getPKH();
         walletInfo.setAddress(userAddress);
-      }
-
-      const url =
-        "https://api.granadanet.tzkt.io/v1/contracts/KT1LFf3MEDg4uZCtYHw4RM5zpuJEvF2NPYsJ/storage";
-      const response = await fetch(url, { method: "GET" });
-
-      const storage = await response.json();
-
-      hounds.clear();
-      for (let i = 0; i < storage.counter; i++) {
-        addHoundToMap(
-          hounds,
-          storage.hounds[i].owner,
-          generateHound(storage.hounds[i].genome, storage.hounds[i].generation)
-        );
       }
     })();
   }, []);
