+++ src/components/pages/Home.tsx
@@ -7,9 +7,8 @@ import "../css/Home.css";
 import { SVGSlantTop, SVGSlantBottom } from "../svgs";
 import { Link } from "react-router-dom";
 import { WalletContext } from "../context/WalletContext";
-import { HoundInfo } from "../../scripts/hound-genome";
 
-function ConnectWallet({ hounds }: { hounds: Map<string, HoundInfo[]> }) {
+function ConnectWallet() {
   const [Tezos] = useState<TezosToolkit>(new TezosToolkit("https://api.tez.ie/rpc/granadanet"));
   const [, setPublicToken] = useState<string | null>("");
   const [isWalletConnected, setWalletConnected] = useState<boolean>(false);
@@ -26,7 +25,6 @@ function ConnectWallet({ hounds }: { hounds: Map<string, HoundInfo[]> }) {
   return (
     <ConnectButton
       Tezos={Tezos}
-      hounds={hounds}
       setPublicToken={setPublicToken}
       setWalletConnected={setWalletConnected}
     />
@@ -45,7 +43,7 @@ function PlayButton() {
   );
 }
 
-function Hero({ hounds }: { hounds: Map<string, HoundInfo[]> }) {
+function Hero() {
   return (
     <div className="hero">
       <div className="hero__text">
@@ -60,7 +58,7 @@ function Hero({ hounds }: { hounds: Map<string, HoundInfo[]> }) {
       </div>
 
       <div className="hero__buttons">
-        <ConnectWallet hounds={hounds} />
+        <ConnectWallet/>
         <PlayButton />
       </div>
     </div>
@@ -135,10 +133,10 @@ function AboutGame() {
   );
 }
 
-function Home({ hounds }: { hounds: Map<string, HoundInfo[]> }): JSX.Element {
+function Home(): JSX.Element {
   return (
     <div>
-      <Hero hounds={hounds} />
+      <Hero />
 
       <SVGSlantTop />
       <AboutGame />
